+++ verify/linalg/euler_circs.test.cpp
@@ -10,7 +10,7 @@ using namespace std;
 using namespace cp_algo::math;
 using namespace cp_algo::linalg;
 
-const int mod = 998244353;
+const int64_t mod = 998244353;
 using base = modint<mod>;
 
 void solve() {
