+++ verify/linalg/prod.test.cpp
@@ -8,7 +8,7 @@ using namespace std;
 using namespace cp_algo::linalg;
 using namespace cp_algo::math;
 
-const int mod = 998244353;
+const int64_t mod = 998244353;
 using base = modint<mod>;
 
 void solve() {
